增加了提醒无法移动的model部分

+++ src/main/java/controller/InterfaceController.java
@@ -130,7 +130,7 @@ public class InterfaceController {
     public void updateGrid() {
         for (int i = 0; i < 4; i++) {
             for (int j = 0; j < 4; j++) {
-                int number = grid.getNumber(i, j);
+                long number = grid.getNumber(i, j);
                 if (number == 0) {
                     labels[i][j].setImage(null);
                 } else {
